+++ packages/images/optimize.d.ts
@@ -1,1 +1,1 @@
-export function svgOptions(): Record<string, unknown>;
+export function svgOptions(): {[key: string]: unknown};